Added validation to the movie rental section.

+++ Vidly/Models/Movies.cs
@@ -27,7 +27,7 @@ namespace Vidly.Models
         [Range(1, 20)]
         public int NumberInStock { get; set; }
 
-        public byte NumberAvailable { get; set; }
+        public int NumberAvailable { get; set; }
 
         [Display(Name = "Genre")]
         [Required]
